+++ include/inc_lib/revision/revision.php
@@ -10,7 +10,7 @@
  **/
 
 define('PHPWCMS_VERSION',		'1.6.522');
-define('PHPWCMS_RELEASE_DATE',	'2012/11/24');
+define('PHPWCMS_RELEASE_DATE',	'2012/11/26');
 define('PHPWCMS_REVISION',		'522');
 
 ?>